% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_excel.R
\name{summary_excel}
\alias{summary_excel}
\title{Summarize Output of extract_excel}
\usage{
summary_excel(data_)
}
\arguments{
\item{data_}{A list of list of dataframes, as returned by \code{\link{extract_excel}}}
}
\value{
A dataframe that summarizes extracted Excel data
}
\description{
Summarize Output of extract_excel
}

