% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_excel.R
\name{view_excel}
\alias{view_excel}
\title{View the Output of extract_excel}
\usage{
view_excel(data_)
}
\arguments{
\item{data_}{A list of list of dataframes, as returned by \code{\link{extract_excel}}}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
A wrapper for `View`
}
\details{
\code{View} all dataframes stored in the output of \code{\link{extract_excel}} at once.
}

